// ws/src/types.ts
export interface DbWritePayload {
    type: 'CREATE_MOVE' | 'UPDATE_GAME_STATUS' | 'ASSIGN_PLAYER' | 'CREATE_MATCHED_GAME';
    payload: any; // Define specific payload structures later
}